Move PlaneProbingEstimatorCommon to geometry/helpers
