+++ src/DGtal/geometry/surfaces/estimation/PlaneProbingNeighborhood.h
@@ -44,7 +44,7 @@
 #include <iostream>
 #include <vector>
 #include "DGtal/base/Common.h"
-#include "DGtal/geometry/surfaces/estimation/PlaneProbingEstimatorCommon.h"
+#include "DGtal/geometry/helpers/PlaneProbingEstimatorHelper.h"
 #include "DGtal/kernel/CPointPredicate.h"
 //////////////////////////////////////////////////////////////////////////////
