+++ src/DGtal/geometry/surfaces/estimation/PlaneProbingTetrahedronEstimator.h
@@ -43,13 +43,49 @@
 #include <iostream>
 #include <vector>
 #include "DGtal/base/Common.h"
-#include "DGtal/geometry/surfaces/estimation/PlaneProbingEstimatorCommon.h"
+#include "DGtal/geometry/helpers/PlaneProbingEstimatorHelper.h"
 #include "DGtal/geometry/surfaces/estimation/PlaneProbingNeighborhood.h"
 #include "DGtal/kernel/CPointPredicate.h"
 //////////////////////////////////////////////////////////////////////////////
 
 namespace DGtal
 {
+    /**
+     * Probing mode for plane-probing estimators.
+     */
+    enum class ProbingMode
+    {
+        H,
+        R,
+        R1,
+    };
+
+    /**
+     * Display a mode on the standard output.
+     *
+     * @param aOs the output stream.
+     * @param aMode the mode to display.
+     */
+    inline
+    std::ostream& operator<< (std::ostream& aOs, ProbingMode const& aMode)
+    {
+        switch (aMode)
+        {
+            case ProbingMode::H:
+                aOs << "H";
+                break;
+
+            case ProbingMode::R:
+                aOs << "R";
+                break;
+
+            case ProbingMode::R1:
+                aOs << "R1";
+                break;
+        }
+
+        return aOs;
+    }
 
   /////////////////////////////////////////////////////////////////////////////
   // template class PlaneProbingTetrahedronEstimator
@@ -152,7 +188,7 @@ namespace DGtal
     Triangle vertices () const;
 
     /**
-     * Get a basis of the lattice generated by the estimated normal.
+     * Get a basis of the lattice generated by the current estimated normal.
      */
     std::pair<Point, Point> getBasis() const;
